// Frontend/src/app/api/models/rh-dashboard-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ApplicationStatsDto } from '../models/application-stats-dto';
import { CommissionStatsDto } from '../models/commission-stats-dto';
import { DashboardAlertDto } from '../models/dashboard-alert-dto';
import { DashboardOverviewDto } from '../models/dashboard-overview-dto';
import { JobOfferStatsDto } from '../models/job-offer-stats-dto';
import { RecentActivityDto } from '../models/recent-activity-dto';
import { TopJobOfferDto } from '../models/top-job-offer-dto';
export interface RhDashboardDto {
  alerts?: Array<DashboardAlertDto>;
  applicationStats?: ApplicationStatsDto;
  commissionStats?: CommissionStatsDto;
  jobOfferStats?: JobOfferStatsDto;
  overview?: DashboardOverviewDto;
  recentActivities?: Array<RecentActivityDto>;
  topJobOffers?: Array<TopJobOfferDto>;
}