// Frontend/src/app/api/models/application-stats-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { JobOfferApplicationCountDto } from '../models/job-offer-application-count-dto';
export interface ApplicationStatsDto {
  applicationsByEducationLevel?: {
[key: string]: number;
};
  applicationsLast30Days?: number;
  averageIAScore?: number;
  growthRate?: number;
  rejectedApplications?: number;
  shortlistedApplications?: number;
  submittedApplications?: number;
  topOffersByApplications?: Array<JobOfferApplicationCountDto>;
  totalApplications?: number;
  underReviewApplications?: number;
  withdrawnApplications?: number;
}