// Frontend/src/app/api/models/job-offer-stats-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface JobOfferStatsDto {
  archivedJobOffers?: number;
  averageCandidatesPerOffer?: number;
  closedJobOffers?: number;
  draftJobOffers?: number;
  jobOffersByContract?: {
[key: string]: number;
};
  jobOffersByType?: {
[key: string]: number;
};
  jobOffersExpiringSoon?: number;
  publishedJobOffers?: number;
  totalJobOffers?: number;
}